Custom tab: extract to WebTools.
